refactor to new file structure for onboarding
