+++ src/onboarding/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  ...require('./handle-new-member'),
+  ...require('./handle-new-message'),
+  ...require('./handle-updated-message'),
+  ...require('./cleanup'),
+}
